experiments/experiment_3/requirements.txt
Flask==2.3.3
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
authlib==1.0.0  # Usar authlib en lugar de Flask-OAuthlib
requests==2.31.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
